+++ abstracttree/__init__.py
@@ -21,13 +21,13 @@ __all__ = [
     "Route",
 ]
 
+from .adapters import astree
 from .binarytree import BinaryTree, BinaryDownTree
-from .conversions import astree
 from .export import print_tree, plot_tree, to_image, to_dot, to_mermaid, to_string, to_pillow, \
     to_latex
 from .heaptree import HeapTree
 from .predicates import RemoveDuplicates, PreventCycles, MaxDepth
 from .route import Route
-from .treeclasses import Tree, DownTree, MutableDownTree, MutableTree
+from .tree import Tree, DownTree, MutableDownTree, MutableTree
 
-__version__ = "0.0.5"
+__version__ = "0.0.6"
